+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## unreleased
+* Add `paypalMessages` to `Dispute`
 * Fix bug where `__isset` methods in `Instance` and `Base` classes treated `null` value as set (Thanks @sklodzio)
 * Add `tax_identifiers` parameter to `Customer.create` and `Customer.update`